refactor(board): rename to board query model
