+++ src/main/java/me/nettee/board/application/usecase/BoardReadUseCase.java
@@ -1,8 +1,8 @@
 package me.nettee.board.application.usecase;
 
-import me.nettee.board.application.model.BoardModel.BoardDetail;
+import me.nettee.board.application.model.BoardQueryModel.BoardDetail;
 
 public interface BoardReadUseCase {
 
-   BoardDetail getBoard(Long id);
+    BoardDetail getBoard(Long id);
 }
